* Delete `StarDustHandler.kt` as it's no longer needed * Add `isPlayerPresent()` function this checks to make sure there no players around the star before starting a new event * Move cleanseEvent() from ShootingStarHandler.kt to ShootingStar.kts * Add objectDespawn for when the star object id equals "shooting_star_tier_1"
